Have a naive working version with benchmarking

+++ Core/Inc/Harris_corner_detection.h
@@ -9,9 +9,9 @@
 
 
 typedef struct{
-	int32_t Ixx;
-	int32_t Ixy;
-	int32_t Iyy;
+	int64_t Ixx;
+	int64_t Ixy;
+	int64_t Iyy;
 }Matrix_values_t;
